refactor: move o conteúdo do app para o seu próprio componente
Tratando todo o conteúdo da tela do app como se fosse uma página
